backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
psycopg[binary]
python-dotenv
pydantic[email]
python-multipart
